Add new common structures for trading, orderbook, and Lightning
