+++ packages/komodo_defi_rpc_methods/lib/src/rpc_methods/orderbook/orderbook_rpc_namespace.dart
@@ -1,10 +1,67 @@
 import 'package:komodo_defi_rpc_methods/komodo_defi_rpc_methods.dart';
 
-/// Extensions for Orderbook-related RPC methods
+/// RPC namespace for orderbook operations.
+/// 
+/// This namespace provides methods for interacting with the decentralized
+/// orderbook in the Komodo DeFi Framework. It enables users to view market
+/// depth, place orders, and manage their trading positions.
+/// 
+/// ## Key Features:
+/// 
+/// - **Market Data**: View orderbook depth and best prices
+/// - **Order Management**: Place, cancel, and monitor orders
+/// - **Price Discovery**: Find the best available prices for trades
+/// - **Order Types**: Support for both maker and taker orders
+/// 
+/// ## Order Lifecycle:
+/// 
+/// 1. **Creation**: Orders are placed using `setOrder`
+/// 2. **Matching**: Orders wait in the orderbook for matching
+/// 3. **Execution**: Matched orders proceed to atomic swap
+/// 4. **Completion**: Orders are removed after execution or cancellation
+/// 
+/// ## Usage Example:
+/// 
+/// ```dart
+/// final orderbook = client.orderbook;
+/// 
+/// // View orderbook
+/// final book = await orderbook.orderbook(
+///   base: 'BTC',
+///   rel: 'KMD',
+/// );
+/// 
+/// // Place an order
+/// final order = await orderbook.setOrder(
+///   base: 'BTC',
+///   rel: 'KMD',
+///   price: '100',
+///   volume: '0.1',
+/// );
+/// ```
 class OrderbookMethodsNamespace extends BaseRpcMethodNamespace {
+  /// Creates a new [OrderbookMethodsNamespace] instance.
+  /// 
+  /// This is typically called internally by the [KomodoDefiRpcMethods] class.
   OrderbookMethodsNamespace(super.client);
 
-  /// Get orderbook for a trading pair
+  /// Retrieves the orderbook for a specific trading pair.
+  /// 
+  /// This method fetches the current state of the orderbook, including
+  /// all active buy and sell orders for the specified trading pair.
+  /// 
+  /// - [base]: The base coin of the trading pair
+  /// - [rel]: The rel/quote coin of the trading pair
+  /// - [rpcPass]: Optional RPC password override
+  /// 
+  /// Returns a [Future] that completes with an [OrderbookResponse]
+  /// containing lists of bids and asks.
+  /// 
+  /// The orderbook includes:
+  /// - **Bids**: Buy orders sorted by price (highest first)
+  /// - **Asks**: Sell orders sorted by price (lowest first)
+  /// - Order details including price, volume, and age
+  /// - Timestamp of the orderbook snapshot
   Future<OrderbookResponse> orderbook({
     required String base,
     required String rel,
@@ -19,7 +76,22 @@ class OrderbookMethodsNamespace extends BaseRpcMethodNamespace {
     );
   }
 
-  /// Get orderbook depth for a trading pair
+  /// Retrieves orderbook depth for multiple trading pairs.
+  /// 
+  /// This method efficiently fetches depth information for multiple
+  /// trading pairs in a single request, useful for market overview
+  /// displays or price aggregation.
+  /// 
+  /// - [pairs]: List of trading pairs to query
+  /// - [rpcPass]: Optional RPC password override
+  /// 
+  /// Returns a [Future] that completes with an [OrderbookDepthResponse]
+  /// containing depth data for each requested pair.
+  /// 
+  /// Depth information includes:
+  /// - Best bid and ask prices
+  /// - Available volume at best prices
+  /// - Number of orders at each price level
   Future<OrderbookDepthResponse> orderbookDepth({
     required List<OrderbookPair> pairs,
     String? rpcPass,
@@ -32,7 +104,24 @@ class OrderbookMethodsNamespace extends BaseRpcMethodNamespace {
     );
   }
 
-  /// Get best orders
+  /// Finds the best orders for a specific trading action.
+  /// 
+  /// This method searches the orderbook to find the best available
+  /// orders that can fulfill a desired trade volume. It's useful for
+  /// determining the expected execution price for market orders.
+  /// 
+  /// - [coin]: The coin to trade
+  /// - [action]: Whether to buy or sell
+  /// - [volume]: The desired trade volume
+  /// - [rpcPass]: Optional RPC password override
+  /// 
+  /// Returns a [Future] that completes with a [BestOrdersResponse]
+  /// containing the best available orders.
+  /// 
+  /// The response includes:
+  /// - Orders sorted by best price
+  /// - Cumulative volume information
+  /// - Average execution price for the volume
   Future<BestOrdersResponse> bestOrders({
     required String coin,
     required OrderType action,
@@ -49,7 +138,33 @@ class OrderbookMethodsNamespace extends BaseRpcMethodNamespace {
     );
   }
 
-  /// Create a new order (maker)
+  /// Places a new maker order on the orderbook.
+  /// 
+  /// This method creates a new limit order that will be added to the
+  /// orderbook and wait for a matching taker. The order remains active
+  /// until it's matched, cancelled, or expires.
+  /// 
+  /// - [base]: The base coin to trade
+  /// - [rel]: The rel/quote coin to trade
+  /// - [price]: The price per unit of base coin in rel coin
+  /// - [volume]: The amount of base coin to trade
+  /// - [orderType]: Optional order type specification
+  /// - [minVolume]: Optional minimum acceptable volume for partial fills
+  /// - [baseConfs]: Optional required confirmations for base coin
+  /// - [baseNota]: Optional nota requirement for base coin
+  /// - [relConfs]: Optional required confirmations for rel coin
+  /// - [relNota]: Optional nota requirement for rel coin
+  /// - [rpcPass]: Optional RPC password override
+  /// 
+  /// Returns a [Future] that completes with a [SetOrderResponse]
+  /// containing the created order details.
+  /// 
+  /// ## Order Configuration:
+  /// 
+  /// - **Price**: Must be positive and reasonable for the market
+  /// - **Volume**: Must exceed minimum trading requirements
+  /// - **Confirmations**: Higher values increase security but slow execution
+  /// - **Nota**: Requires notarization for additional security
   Future<SetOrderResponse> setOrder({
     required String base,
     required String rel,
@@ -80,7 +195,19 @@ class OrderbookMethodsNamespace extends BaseRpcMethodNamespace {
     );
   }
 
-  /// Cancel an order
+  /// Cancels a specific order.
+  /// 
+  /// This method removes an active order from the orderbook. Only orders
+  /// that haven't been matched can be cancelled.
+  /// 
+  /// - [uuid]: The unique identifier of the order to cancel
+  /// - [rpcPass]: Optional RPC password override
+  /// 
+  /// Returns a [Future] that completes with a [CancelOrderResponse]
+  /// indicating the cancellation result.
+  /// 
+  /// Note: Orders that are already matched and proceeding to swap
+  /// cannot be cancelled.
   Future<CancelOrderResponse> cancelOrder({
     required String uuid,
     String? rpcPass,
@@ -93,7 +220,26 @@ class OrderbookMethodsNamespace extends BaseRpcMethodNamespace {
     );
   }
 
-  /// Cancel all orders for a specific coin or all orders
+  /// Cancels multiple orders based on the specified criteria.
+  /// 
+  /// This method provides bulk cancellation functionality, allowing users
+  /// to cancel all their orders or all orders for a specific coin.
+  /// 
+  /// - [cancelType]: Specifies which orders to cancel (all or by coin)
+  /// - [rpcPass]: Optional RPC password override
+  /// 
+  /// Returns a [Future] that completes with a [CancelAllOrdersResponse]
+  /// containing the results of the cancellation operation.
+  /// 
+  /// ## Cancel Types:
+  /// 
+  /// - `CancelOrdersType.all()`: Cancels all active orders
+  /// - `CancelOrdersType.coin("BTC")`: Cancels all orders involving BTC
+  /// 
+  /// This is useful for:
+  /// - Emergency stops
+  /// - Portfolio rebalancing
+  /// - Market exit strategies
   Future<CancelAllOrdersResponse> cancelAllOrders({
     CancelOrdersType? cancelType,
     String? rpcPass,
@@ -106,7 +252,21 @@ class OrderbookMethodsNamespace extends BaseRpcMethodNamespace {
     );
   }
 
-  /// Get my orders
+  /// Retrieves all orders created by the current user.
+  /// 
+  /// This method returns a comprehensive list of the user's orders,
+  /// including their current status and match information.
+  /// 
+  /// - [rpcPass]: Optional RPC password override
+  /// 
+  /// Returns a [Future] that completes with a [MyOrdersResponse]
+  /// containing all user orders.
+  /// 
+  /// The response includes:
+  /// - Active orders waiting for matches
+  /// - Orders currently being matched
+  /// - Recently completed or cancelled orders
+  /// - Detailed status and configuration for each order
   Future<MyOrdersResponse> myOrders({
     String? rpcPass,
   }) {
