+++ packages/komodo_defi_rpc_methods/lib/src/rpc_methods/trading/trading_rpc_namespace.dart
@@ -1,10 +1,70 @@
 import 'package:komodo_defi_rpc_methods/komodo_defi_rpc_methods.dart';
 
-/// Extensions for Trading/Swap-related RPC methods
+/// RPC namespace for trading and swap operations.
+/// 
+/// This namespace provides methods for managing atomic swaps and trading
+/// operations within the Komodo DeFi Framework. It enables users to initiate,
+/// monitor, and manage cross-chain atomic swaps in a decentralized manner.
+/// 
+/// ## Key Features:
+/// 
+/// - **Swap Initiation**: Start new swaps as maker or taker
+/// - **Swap Monitoring**: Track active and recent swap status
+/// - **Trade Analysis**: Calculate fees and validate trade parameters
+/// - **Swap Management**: Cancel active swaps when needed
+/// 
+/// ## Swap Types:
+/// 
+/// - **Maker**: Sets an order at a specific price and waits for takers
+/// - **Taker**: Takes existing orders from the orderbook immediately
+/// 
+/// ## Usage Example:
+/// 
+/// ```dart
+/// final trading = client.trading;
+/// 
+/// // Start a new swap
+/// final swap = await trading.startSwap(
+///   swapRequest: SwapRequest(
+///     base: 'BTC',
+///     rel: 'KMD',
+///     baseCoinAmount: '0.1',
+///     relCoinAmount: '1000',
+///     method: SwapMethod.sell,
+///   ),
+/// );
+/// 
+/// // Monitor swap status
+/// final status = await trading.swapStatus(uuid: swap.uuid);
+/// ```
 class TradingMethodsNamespace extends BaseRpcMethodNamespace {
+  /// Creates a new [TradingMethodsNamespace] instance.
+  /// 
+  /// This is typically called internally by the [KomodoDefiRpcMethods] class.
   TradingMethodsNamespace(super.client);
 
-  /// Start a new swap
+  /// Initiates a new atomic swap.
+  /// 
+  /// This method starts a new cross-chain atomic swap based on the provided
+  /// parameters. The swap can be initiated as either a maker (placing an order)
+  /// or a taker (taking an existing order).
+  /// 
+  /// - [swapRequest]: The swap configuration including coins, amounts, and method
+  /// - [rpcPass]: Optional RPC password override
+  /// 
+  /// Returns a [Future] that completes with a [StartSwapResponse] containing
+  /// the swap UUID and initial status.
+  /// 
+  /// ## Swap Methods:
+  /// 
+  /// - **setPrice**: Creates a maker order at a specific price
+  /// - **buy**: Takes the best available sell orders (taker)
+  /// - **sell**: Takes the best available buy orders (taker)
+  /// 
+  /// Throws an exception if:
+  /// - Insufficient balance for the swap
+  /// - No matching orders available (for taker swaps)
+  /// - Invalid swap parameters
   Future<StartSwapResponse> startSwap({
     required SwapRequest swapRequest,
     String? rpcPass,
@@ -17,7 +77,22 @@ class TradingMethodsNamespace extends BaseRpcMethodNamespace {
     );
   }
 
-  /// Get swap status
+  /// Retrieves the status of a specific swap.
+  /// 
+  /// This method fetches detailed information about a swap identified by
+  /// its UUID, including current state, progress, and transaction details.
+  /// 
+  /// - [uuid]: The unique identifier of the swap
+  /// - [rpcPass]: Optional RPC password override
+  /// 
+  /// Returns a [Future] that completes with a [SwapStatusResponse] containing
+  /// comprehensive swap information.
+  /// 
+  /// The status includes:
+  /// - Current swap state and progress
+  /// - Transaction IDs and confirmations
+  /// - Error information if the swap failed
+  /// - Timestamps for each swap event
   Future<SwapStatusResponse> swapStatus({
     required String uuid,
     String? rpcPass,
@@ -30,7 +105,22 @@ class TradingMethodsNamespace extends BaseRpcMethodNamespace {
     );
   }
 
-  /// Get active swaps
+  /// Retrieves all currently active swaps.
+  /// 
+  /// This method returns information about all swaps that are currently
+  /// in progress. Optionally, results can be filtered by coin.
+  /// 
+  /// - [coin]: Optional coin ticker to filter results
+  /// - [rpcPass]: Optional RPC password override
+  /// 
+  /// Returns a [Future] that completes with an [ActiveSwapsResponse]
+  /// containing lists of active swap UUIDs and their details.
+  /// 
+  /// Active swaps include those that are:
+  /// - Waiting for maker payment
+  /// - Waiting for taker payment
+  /// - Waiting for confirmations
+  /// - In any other non-terminal state
   Future<ActiveSwapsResponse> activeSwaps({
     String? coin,
     String? rpcPass,
@@ -43,7 +133,23 @@ class TradingMethodsNamespace extends BaseRpcMethodNamespace {
     );
   }
 
-  /// Get recent swaps with pagination
+  /// Retrieves recent swap history with pagination support.
+  /// 
+  /// This method fetches historical swap data, including both completed
+  /// and failed swaps. Results can be paginated and filtered by coin.
+  /// 
+  /// - [limit]: Maximum number of swaps to return
+  /// - [fromUuid]: Starting point for pagination (exclusive)
+  /// - [coin]: Optional coin ticker to filter results
+  /// - [rpcPass]: Optional RPC password override
+  /// 
+  /// Returns a [Future] that completes with a [RecentSwapsResponse]
+  /// containing swap history records.
+  /// 
+  /// ## Pagination:
+  /// 
+  /// To paginate through results, use the UUID of the last swap from
+  /// the previous response as the [fromUuid] parameter.
   Future<RecentSwapsResponse> recentSwaps({
     int? limit,
     int? fromUuid,
@@ -60,7 +166,20 @@ class TradingMethodsNamespace extends BaseRpcMethodNamespace {
     );
   }
 
-  /// Cancel an active swap
+  /// Cancels an active swap.
+  /// 
+  /// This method attempts to cancel a swap that is currently in progress.
+  /// Cancellation is only possible for swaps in certain states, typically
+  /// before the payment transactions have been broadcast.
+  /// 
+  /// - [uuid]: The unique identifier of the swap to cancel
+  /// - [rpcPass]: Optional RPC password override
+  /// 
+  /// Returns a [Future] that completes with a [CancelSwapResponse]
+  /// indicating whether the cancellation was successful.
+  /// 
+  /// Note: Swaps cannot be cancelled after payment transactions have
+  /// been broadcast to prevent loss of funds.
   Future<CancelSwapResponse> cancelSwap({
     required String uuid,
     String? rpcPass,
@@ -73,7 +192,27 @@ class TradingMethodsNamespace extends BaseRpcMethodNamespace {
     );
   }
 
-  /// Get trade preimage
+  /// Calculates fees and validates parameters for a potential trade.
+  /// 
+  /// This method performs a dry-run calculation of a trade, providing
+  /// fee estimates and validation without actually initiating the swap.
+  /// It's useful for showing users the expected costs before confirmation.
+  /// 
+  /// - [base]: The base coin ticker
+  /// - [rel]: The rel/quote coin ticker
+  /// - [swapMethod]: The intended swap method (setPrice, buy, or sell)
+  /// - [volume]: The trade volume
+  /// - [price]: Optional price for maker orders
+  /// - [rpcPass]: Optional RPC password override
+  /// 
+  /// Returns a [Future] that completes with a [TradePreimageResponse]
+  /// containing fee calculations and validation results.
+  /// 
+  /// The preimage includes:
+  /// - Estimated transaction fees for both coins
+  /// - Actual tradeable volume after fees
+  /// - Validation of trade parameters
+  /// - Required transaction confirmations
   Future<TradePreimageResponse> tradePreimage({
     required String base,
     required String rel,
@@ -94,7 +233,22 @@ class TradingMethodsNamespace extends BaseRpcMethodNamespace {
     );
   }
 
-  /// Get max taker volume
+  /// Calculates the maximum volume available for a taker swap.
+  /// 
+  /// This method determines the maximum amount of a coin that can be
+  /// traded as a taker, considering available balance and required fees.
+  /// 
+  /// - [coin]: The coin ticker to check
+  /// - [rpcPass]: Optional RPC password override
+  /// 
+  /// Returns a [Future] that completes with a [MaxTakerVolumeResponse]
+  /// containing the maximum tradeable volume.
+  /// 
+  /// The calculation considers:
+  /// - Available coin balance
+  /// - Required transaction fees
+  /// - Dust limits
+  /// - Protocol-specific constraints
   Future<MaxTakerVolumeResponse> maxTakerVolume({
     required String coin,
     String? rpcPass,
@@ -107,7 +261,21 @@ class TradingMethodsNamespace extends BaseRpcMethodNamespace {
     );
   }
 
-  /// Get min trading volume
+  /// Retrieves the minimum trading volume for a coin.
+  /// 
+  /// This method returns the minimum amount of a coin that can be
+  /// traded in a swap, considering dust limits and economic viability.
+  /// 
+  /// - [coin]: The coin ticker to check
+  /// - [rpcPass]: Optional RPC password override
+  /// 
+  /// Returns a [Future] that completes with a [MinTradingVolumeResponse]
+  /// containing the minimum tradeable amount.
+  /// 
+  /// The minimum is determined by:
+  /// - Protocol dust limits
+  /// - Transaction fee requirements
+  /// - Economic viability thresholds
   Future<MinTradingVolumeResponse> minTradingVolume({
     required String coin,
     String? rpcPass,
